requirements.txt
ansitable==0.11.3
cfgv==3.4.0
colored==2.2.4
cycler==0.12.1
distlib==0.3.9
filelock==3.16.1
fonttools==4.54.1
identify==2.6.1
kiwisolver==1.4.7
matplotlib==3.5.1
nodeenv==1.9.1
numpy==1.26.4
packaging==24.2
pgraph-python==0.6.2
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
progress==1.6
pybullet==3.2.6
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qpsolvers==2.3.0
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scipy==1.14.1
six==1.16.0
spatialgeometry==1.1.0
spatialmath-python==1.1.13
swift-sim==1.1.0
typing_extensions==4.12.2
virtualenv==20.27.1
websockets==13.1
xacro==1.13.3
